requirements.txt
streamlit
openai
jinja2
requests
python-dotenv
loguru
